Use `setup_test_component_platform` helper for select entity component tests instead of `hass.components`
