+++ tests/components/select/test_init.py
@@ -21,6 +21,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ServiceValidationError
 from homeassistant.setup import async_setup_component
 
+from tests.common import setup_test_component_platform
+
 
 class MockSelectEntity(SelectEntity):
     """Mock SelectEntity to use in tests."""
@@ -91,11 +93,11 @@ async def test_select(hass: HomeAssistant) -> None:
 
 
 async def test_custom_integration_and_validation(
-    hass: HomeAssistant, enable_custom_integrations: None
+    hass: HomeAssistant,
+    mock_select_entities: list[MockSelectEntity],
 ) -> None:
     """Test we can only select valid options."""
-    platform = getattr(hass.components, f"test.{DOMAIN}")
-    platform.init()
+    setup_test_component_platform(hass, DOMAIN, mock_select_entities)
 
     assert await async_setup_component(hass, DOMAIN, {DOMAIN: {CONF_PLATFORM: "test"}})
     await hass.async_block_till_done()
